Use RDRAND in addition to urandom.
Through some discussions it became apparent alternative security models
are held by some rather motivated community members.  Out of a desire to
minimize community fragmentation this patch combines the entropy from
RDRAND and urandom.

Recall the intent of entropy is not speed but to seed fast and closely
held DRBGs.  Hence, securely combining N sources should be acceptable to
user or they are actually looking for a slightly different package with
different goals.