Separate system and HW entropy sources.
Since this was once a hot topic this commit deserves a few comments.

Some users are resistant to the inclusion of RDRAND, most notably issue #30
but also the unfortunately-heated issue #14 (now ancient). This is
because the ease with which a backdoor could live inside any hardware
RNG and how especially damaging it can be if able to peek at the inputs
of other sources.

My stance - prior and present - has been that software back-doors (such
as one impacting `urandom`) are of greater concern wrt mass-surveillance
or APTs.  I'd even cynically say people are unintentionally trusting UPS
and their front porch more than Intel.  What has changed, aside from my
willingness to spend time on this while waiting for a delayed flight, is
1. I had agreed that the entropy package is supposed to be lower-level
in that it provides the most fundamental primitives of access to the
system _and_ the hardware RNG as rather blatantly evidenced by the very
module name it occupies and 2. I have a long-running concern that
eventually someone disgruntled will fork entropy, rather than pursue any
of the discussed alternatives for which I'd accept a patch, and thus
cause more fragmentation and package clutter for upstream devs.  The
community is already fragile in too many regards.

Let's reflect on the first issue - that `entropy` and any module claiming
the name `System.Random` is low level and should provide the most
primitive operations.  It would be appropriate to develop a light weight
and formally analyzed randomness pool such as Fortuna.  Entropy could be
a dependency of this library.  I acknowledge the benefit to complexity
ratio is speculative and require buy-in to a mental model of
the adversary that, discussions have revealed, is unpopular.

-TomMD