Add getrandom() entropy source
For old glibc missing the getrandom() function, there is a fall-back
that directly calls the getrandom system call. If getrandom() cannot be
used it will fall back to /dev/urandom, but with an additional check to
make sure that the entropy pool has been initialized. Without this
check, if an application is run during boot then /dev/urandom might not
be a cryptographically secure random number generator.